booking/serializers/profile.py
from rest_framework import serializers
from booking.models import Profile


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = ['phone_number']
